+++ src/components/settings/PullsSettingsApp.jsx
@@ -1,37 +1,10 @@
 const { ipcRenderer } = require('electron');
-const definedSettings = require('./settings-config.js');
 const settings = require('./settings-utils.js');
-
-
-const path = require('path');
 
 import swal from 'sweetalert';
 
-import React from 'react';
-import SettingInput from './SettingInput';
-import { useSettings } from '../../hooks/useSettings';
-
-import Editor from "@monaco-editor/react";
-import { loader } from "@monaco-editor/react";
-
-function ensureFirstBackSlash(str) {
-    return str.length > 0 && str.charAt(0) !== "/"
-        ? "/" + str
-        : str;
-}
-
-function uriFromPath(_path) {
-    const pathName = path.resolve(_path).replace(/\\/g, "/");
-    return encodeURI("file://" + ensureFirstBackSlash(pathName));
-}
-
-loader.config({
-  paths: {
-    vs: uriFromPath(
-      path.join(__dirname, "../node_modules/monaco-editor/min/vs")
-    )
-  }
-});
+import React, { useState } from 'react';
+import PreferencesEditor from './PreferencesEditor.jsx';
 
 function PullsSettingsApp() {
     function _handleSave() {
@@ -63,27 +36,9 @@ function PullsSettingsApp() {
             ></i>
         </div>
 
-        <Editor
-            height="90vh"
-            defaultLanguage="json"
-            options={{minimap: { enabled: false }}}
-            defaultValue={JSON.stringify({
-                githubUser: 'matthewnitschke-wk',
-                queries: [
-                    { label: 'My Prs', query: 'is:pr' }
-                ]
-            }, null, 4)}
-            theme='vs-dark'
-        />
-
-        {/* <div className="settings-editor-app-body">
-            {definedSettings.map(settingOptions => {
-                return <SettingInput
-                    key={settingOptions.settingsKey}
-                    {...settingOptions}
-                    onSave={_handleSave} />
-            })}
-        </div> */}
+        <PreferencesEditor
+            defaultValue={JSON.stringify(settings.get('preferences'), null, 2)}
+            onChange={val => settings.set('preferences', JSON.parse(val))} />
     </div>
 }
